% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_data.R
\name{country_names}
\alias{country_names}
\title{Get list of available country names from rnaturalearth}
\usage{
country_names(scale = 110)
}
\arguments{
\item{scale}{of map to use, one of 110, 50, 10}
}
\description{
Get list of available country names from rnaturalearth
}
\note{
At \code{scale = 110}, 177 countries are available. At \code{scale=50}, 241 countries are available. At \code{scale=10}, 255 countries are available.
}
