% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlwidget.R
\name{geovis-shiny}
\alias{geovis-shiny}
\alias{geovisOutput}
\alias{renderGeovis}
\title{Shiny bindings for geovis}
\usage{
geovisOutput(output_id, width = "100\%", height = "400px")

renderGeovis(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{output_id}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a geovis}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using geovis within Shiny
applications and interactive Rmd documents.
}
