% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geovis.R
\name{geovis}
\alias{geovis}
\title{Generate a geovis output}
\usage{
geovis(width = NULL, height = NULL)
}
\arguments{
\item{width}{width of output}

\item{height}{height of output}
}
\description{
Generate a geovis output
}
