% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlwidget.R
\name{geovis_widget}
\alias{geovis_widget}
\title{Generate a geovis widget output}
\usage{
geovis_widget(width = NULL, height = NULL, www_dir)
}
\arguments{
\item{width}{width of output}

\item{height}{height of output}

\item{www_dir}{directory where widget will be placed}
}
\description{
Generate a geovis widget output
}
