% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_data.R
\name{get_geo_data}
\alias{get_geo_data}
\title{Get geographic data for a geovis visualization}
\usage{
get_geo_data(countries, get_states = TRUE, get_munis = FALSE,
  scale = 110)
}
\arguments{
\item{countries}{a vector of country names (see \code{\link{country_names}}) for a valid list}

\item{get_states}{should state boundaries within each \code{country} also be retrieved?}

\item{get_munis}{should municipality (county/province) boundaries be retrieved?}

\item{scale}{scale to use when retrieving}
}
\description{
Get geographic data for a geovis visualization
}
\details{
This function uses rnaturalearth to obtain country and state boundary data for a given vector of country names or codes
}
