% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geovis.R
\name{geovis}
\alias{geovis}
\title{Create a geovis visualization}
\usage{
geovis(geo, path = NULL, country_data = NULL, state_data = NULL,
  muni_data = NULL, var_info = NULL)
}
\arguments{
\item{geo}{an object returned from \code{\link{get_geo_data}}}

\item{path}{where to place the widget}

\item{country_data}{data to use at the country level}

\item{state_data}{data to use at the state level}

\item{muni_data}{data to use at the municipality level}

\item{var_info}{information about variables (TODO)}
}
\description{
Create a geovis visualization
}
